transcription-service/requirements.txt
# QuickFlow Transcription Service
# Python 3.10+ required

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
pyannote.audio==3.1.1

# Monitoring
prometheus-client==0.19.0

# Utilities
pydantic>=2.0.0
python-dotenv==1.0.0

# Development
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.26.0
